Don't open ports for getPortNames()
Open certain devices, such as serial endpoints over Bluetooth, can take a long time before they fail. Moreover, it's better to show devices even when the user currently doesn't have access to them - otherwise one might think of a problem with the device or the driver, when a "sudo chmod 666" would be sufficient.